requirements.txt
asgiref==3.2.10
certifi==2024.2.2
charset-normalizer==3.3.2
Django==3.1
gunicorn==21.2.0
idna==3.6
packaging==23.2
python-environ==0.4.54
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.2.1
